feat: Dependencia Swagger adicionada ao pom.mxl fix: ClienteController teve logica alterada apenas para mostrar os resultados das chamadas correto. Antes, a classe estava verificando se  infgormação passava ou nao pelos requisitos básicos. ClienteService agora faz a triagem da informação com lógica de negócios antes de joga-la para o repository.

+++ ProjetoAccentureAcademy/petiscai/src/main/java/acc/br/petiscai/service/ClienteService.java
@@ -4,6 +4,8 @@ import acc.br.petiscai.dto.ClienteDto;
 import acc.br.petiscai.entity.Cliente;
 import acc.br.petiscai.repository.ClienteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,24 +17,33 @@ public class ClienteService {
     ClienteRepository clienteRepository;
 
     public String save(ClienteDto clienteDto) {
-        try {
-            Cliente cliente = new Cliente(clienteDto.getCpf(), clienteDto.getNome(), clienteDto.getEmail(),
-                    clienteDto.getTelefone(), clienteDto.getEndereco(), clienteDto.getIdade());
-            this.clienteRepository.save(cliente);
-            return "Cliente salvo com sucesso!";
-        } catch (Exception e) {
-            return e.getMessage();
+        if (checaIdade(clienteDto)) {
+            try {
+                Cliente cliente = new Cliente(clienteDto.getCpf(), clienteDto.getNome(), clienteDto.getEmail(),
+                        clienteDto.getTelefone(), clienteDto.getEndereco(), clienteDto.getIdade());
+                this.clienteRepository.save(cliente);
+                return "Cliente salvo com sucesso!";
+            } catch (Exception e) {
+                return e.getMessage();
+            }
+        } else {
+            return ("Cliente menor de idade!");
         }
-
     }
 
     public String deleteById(long id) {
-        try {
-            this.clienteRepository.deleteById(id);
-            return "Cliente deletado com sucesso!";
-        } catch (Exception e) {
-            return e.getMessage();
+        if (findByIdBool(id)) {
+
+            try {
+                this.clienteRepository.deleteById(id);
+                return "Cliente deletado com sucesso!";
+            } catch (Exception e) {
+                return e.getMessage();
+            }
+        } else {
+            return "Cliente não encontrado";
         }
+
     }
 
     public List<Cliente> findAll() {
@@ -48,20 +59,34 @@ public class ClienteService {
     }
 
     public Cliente findById(long id) {
-        return this.clienteRepository.findById(id).get();
+
+        if (findByIdBool(id)) {
+            try {
+                return this.clienteRepository.findById(id).get();
+            }catch(Exception e) {
+                return null;
+            }
+        } else {
+            return null;
+        }
     }
+
     public boolean findByIdBool(long id) {
         return this.clienteRepository.findById(id).isPresent();
     }
 
     public String update(ClienteDto clienteDto, long id) {
-        try {
-            Cliente cliente = new Cliente(id, clienteDto.getCpf(), clienteDto.getNome(), clienteDto.getEmail(),
-                    clienteDto.getTelefone(), clienteDto.getEndereco(), clienteDto.getIdade());
-            this.clienteRepository.save(cliente);
-            return "Cliente "+ cliente.getId() + " alterado com sucesso!";
-        } catch (Exception e) {
-            return e.getMessage();
+        if (findByIdBool(id)) {
+            try {
+                Cliente cliente = new Cliente(id, clienteDto.getCpf(), clienteDto.getNome(), clienteDto.getEmail(),
+                        clienteDto.getTelefone(), clienteDto.getEndereco(), clienteDto.getIdade());
+                this.clienteRepository.save(cliente);
+                return "Cliente " + cliente.getId() + " alterado com sucesso!";
+            } catch (Exception e) {
+                return e.getMessage();
+            }
+        } else {
+            return "Cliente não encontrado";
         }
     }
